requirements.txt
click >= 7, < 8
ipython >= 7
ipdb >= 0.11
natsort >= 5
netcdf4 >= 1.2.2
numpy >= 1.15, < 1.20
pandas >= 1.2, < 1.3
pyomo >= 6.0, < 6.1
ruamel.yaml >= 0.16
scikit-learn >= 0.22, < 0.24
xarray >= 0.17, < 0.18
boltons
glom>=19.10
networkx
pydantic
pyyaml
typing_extensions
